glx/glvnd: rewrite auto-generated file
While this file seems to be auto-generated, the script(s) used are not
available. Additionally we've done a ton of work in here already so
might as well make it more comfortable to read.

 - Cleanup unneeded whitespace
 - Use glx.h/glxext.h function typedefs
 - Add macro to manage the func ptr retrieval
 - Flatten the conditionals
 - Remove unneeded variables/variable initialisation.

No functional changes.

Note: this introduces "XXX discards const qualifier" as the header
declaration of GLX_SGIX_fbconfig functions lack the const qualifier for
the attrib_list. This should be fixed upstream (Khornos) who will
release updated/corrected headers.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>