meson: bump meson version requirements to 0.43
There are two features that have made it into 0.43 that make the meson
build system much nicer to work with. One involves picking the C++
linker when linking with a static library that was linked with a static
library containing C++ code instead of the C linker; the other is the
ability to index into CustomTarget's to only include some of their
sources.

This allows us to not just delete code, but also to remove a dependency
on touch, and delete code. For the CustomTarget indexing it also makes
the build system more robust, since the workaround used would not work
if the generated source was not in the same directory as the consumer.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>